Move issues tasks out of folder
